+++ jeap-deploymentlog-docgen/src/main/java/ch/admin/bit/jeap/deploymentlog/docgen/model/DeploymentHistoryOverviewPageDto.java
@@ -18,8 +18,4 @@ public class DeploymentHistoryOverviewPageDto {
         return "Deployment History Overview " + environmentName;
     }
 
-    public String getPageTitle() {
-        return pageTitle(environmentName);
-    }
-
 }