+++ jeap-deploymentlog-docgen/src/main/java/ch/admin/bit/jeap/deploymentlog/docgen/model/DeploymentHistoryPageDto.java
@@ -17,8 +17,4 @@ public class DeploymentHistoryPageDto {
     public static String pageTitle(String systemName, String environmentName) {
         return "Deployment History " + environmentName + " (" + systemName + ")";
     }
-
-    public String getPageTitle() {
-        return pageTitle(systemName, environmentName);
-    }
 }